% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{calc_ss}
\alias{calc_ss}
\title{Calculate within cluster sum of squares}
\usage{
calc_ss(clustdf, distobj, clustcol = "clust")
}
\arguments{
\item{clustdf}{Dataframe with column of cluster membership.}

\item{distobj}{Distance object from which clusters were generated.}

\item{clustcol}{Character. Name of column in clustdf that contains the clusters.}
}
\value{
Dataframe of within group sum-of-squares for each cluster. If it
doesn't already exist in global workspace, sqDist is added there as it is
needed for calculations.
}
\description{
Calculate within cluster sum of squares
}
