% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{clusters_with_freq_taxa}
\alias{clusters_with_freq_taxa}
\title{Clusters with >= 1 taxa with frequency > than threshold}
\usage{
clusters_with_freq_taxa(
  clust_df,
  flor_df,
  clust_col = "clust",
  taxa_col = "taxa",
  site_col = "cell",
  thresh = 0.9
)
}
\arguments{
\item{clust_df}{Dataframe with column of cluster membership and join column to
flordf}

\item{flor_df}{Dataframe with taxa}

\item{clust_col}{Character. Name of column in clustdf with clusters.}

\item{taxa_col}{Character. Name of column in flordf with taxa.}

\item{site_col}{Character. Name of column in clustdf and flordf with 'sites'.}

\item{thresh}{Numeric. Proportion of sites in a cluster at which >= 1 taxa
need to occur.}
}
\value{
Numeric. Number of clusters across clustering that have at least one
taxa that occurs in greater than 'thresh' proportion of sites.
}
\description{
Clusters with >= 1 taxa with frequency > than threshold
}
