% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{clusters_with_indicator}
\alias{clusters_with_indicator}
\title{Clusters with indicator}
\usage{
clusters_with_indicator(indvaldf, thresh = 0.05)
}
\arguments{
\item{indvaldf}{Dataframe from make_indval_df}

\item{thresh}{Numeric. pval threshold for acceptance as indicator taxa.}
}
\value{
Numeric. Number of clusters with at least one indicator taxa.
}
\description{
Clusters with indicator
}
