% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_df.R
\name{diagnostic_df}
\alias{diagnostic_df}
\title{Use a set of (continuous) columns to choose a good set of candidate rows}
\usage{
diagnostic_df(
  df,
  diagnosticdf = tibble::tibble(diagnostic = "avClustSize", highGood = TRUE,
    weightClustSum = TRUE),
  context = c("method", "groups"),
  diagcol = "weightClustSum",
  summarisemethod = median,
  mingroups = 10,
  maxgroups = Inf,
  topthresh = 0.25,
  bestthresh = 5
)
}
\arguments{
\item{df}{Dataframe with columns over which to find good candidates.}

\item{diagnosticdf}{Dataframe mapping the name of possible diagnostics to cases (columns) in which to use those diagnostic.}

\item{context}{Character. Name of columns in df that define context.}

\item{diagcol}{Character. Name of diagnosticdf column to use in this instance.}

\item{summarisemethod}{Character. Name of method to use in summarising if there is more than one row per context.}

\item{mingroups}{Numeric. Minimum number of clusters in a clustering.}

\item{maxgroups}{Numeric. Maximum number of clusters in a clustering.}

\item{topthresh}{Numeric specifying the proportion of rows considered 'top'.}

\item{bestthresh}{Numeric specifying the absolute number of rows considered 'best'.}
}
\value{

}
\description{
Use a set of (continuous) columns to choose a good set of candidate rows
}
