% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_cluster_df}
\alias{make_cluster_df}
\title{Make a cluster data frame}
\usage{
make_cluster_df(raw_clusters, context_df, context)
}
\arguments{
\item{raw_clusters}{Dataframe of cluster membership. Can be result of
\code{make_clusters()}.}

\item{context_df}{Dataframe with \code{context} columns in same order as
\code{raw_clusters}.}

\item{context}{Character. Name of columns in \code{site_df} defining the context.}
}
\value{
Dataframe with 'site','clust' (numeric) and 'cluster' (character)
}
\description{
Make a cluster data frame
}
