% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cluster_sil_mean.R
\name{make_cluster_sil_mean}
\alias{make_cluster_sil_mean}
\title{Mean silhouette width for a cluster}
\usage{
make_cluster_sil_mean(cluster, sil_df)
}
\arguments{
\item{cluster}{Name of cluster}

\item{sil_df}{Output from \code{make_sil_df()}}
}
\value{
Numeric. Mean of silhouette widths for sites in a cluster.
}
\description{
Mean silhouette width for a cluster
}
