% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters_explore.R
\name{make_clusters_explore}
\alias{make_clusters_explore}
\title{Explore clusters}
\usage{
make_clusters_explore(clusters_df, cores = 1, obj_list)
}
\arguments{
\item{clusters_df}{Dataframe with column 'clusters'.}

\item{cores}{How many cores to use in \code{multidplyr::new_cluster()}?}

\item{obj_list}{List of objects required:
\describe{
\item{out_dir}{directory into which to save results.}
\item{dist_flor}{\code{dist(flor_wide)}.}
\item{dist_flor_mat}{\code{as.matrix(dist(flor_wide))}.}
\item{dist_env}{\code{dist(env_wide)}.}
\item{dist_env_mat}{\code{as.matrix(dist(env_wide))}.}
\item{flor_wide}{Wide version of flor_tidy.}
\item{flor_tidy}{Cleaned data set of taxa observations.}
\item{visit_cols}{context.}
\item{p_thresh}{usually 0.05.}
\item{n_sites}{How many unique contexts in \code{flor_tidy}.}
\item{most_freq_prop_thresh}{Threshold for proportion of sites in a cluster that must all contain at least one taxa in common.}
}}
}
\value{
clusters_df with added columns. Saves each output along the way to \code{out_dir}.
}
\description{
Explore clusters using silhouette width, within-group sum-of-squares,
indicators and widespread taxa.
}
