% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dend.R
\name{make_dend}
\alias{make_dend}
\title{Make a dendogram}
\usage{
make_dend(method, dist_bio, dist_env = NULL, geo_alpha = 0.1)
}
\arguments{
\item{method}{What method to use.}

\item{dist_bio}{Distance object, from, say, \code{vegan::vegdist()},
\code{stats::dist()} or \code{parallelDist::parDist()}.}

\item{dist_env}{Distance object. Usually environmental distance. Needs to be
for the same sites as used to create \code{dist_bio}. Only needed if method is
"geo".}

\item{geo_alpha}{Numeric. Value above 0 and below 1 used as \code{alpha} argument
to \code{ClustGeo::hclustgeo()}.}
}
\description{
Make a dendogram
}
