% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gap_df.R
\name{make_gap_df}
\alias{make_gap_df}
\title{Calculate gap statistic for a clustering}
\usage{
make_gap_df(clust_df, dist_mat, clust_col = "cluster", n_sample = 10)
}
\arguments{
\item{clust_df}{Dataframe with context columns and a column with cluster
membership for that context. Optional if \code{clust_col} appears in bio_df.}

\item{dist_mat}{Distance matrix (not distance object)}

\item{clust_col}{Character. Name of column in \code{clusters} identifying cluster
membership.}

\item{n_sample}{Numeric. Number of times to shuffle the membership to
recalculate the wss}
}
\value{
Dataframe with one row per level of \code{clust_col}, summarising the gap
and wss data for each level.
}
\description{
Calculate gap statistic for a clustering
}
