% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_indval_df}
\alias{make_indval_df}
\title{Create a dataframe of indicator values}
\usage{
make_indval_df(clustdf, dfwide, clustcol = "clust")
}
\arguments{
\item{clustdf}{Dataframe including a column with cluster membership}

\item{dfwide}{Dataframe of sites * taxa. In the same order as clustdf.}

\item{clustcol}{Column containing cluster membership in clustdf as name or
index.}
}
\value{
Dataframe of each taxa and the cluster (clust as numeric, cluster as
character) class to which it is most likely an indicator, plus the following
values from labdsv::indval output: indval, pval, abu and frq.
}
\description{
Create a dataframe of indicator values
}
