% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_neigh_text.R
\name{make_neigh_text}
\alias{make_neigh_text}
\title{Make text to describe silhouette.}
\usage{
make_neigh_text(this_clust, sil_df, verbose = TRUE)
}
\arguments{
\item{this_clust}{Name of cluster.}

\item{sil_df}{Output from \code{make_sil_df()}}

\item{verbose}{Logical; return as single sentence (default) or character vector}
}
\value{
Sentence for use in silhouette section of report, or a character vector if \code{verbose=FALSE}
}
\description{
Make text to describe silhouette.
}
