% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sil_df.R
\name{make_sil_df}
\alias{make_sil_df}
\title{Create a dataframe of silhouette widths for a clustering}
\usage{
make_sil_df(clust_df, dist_obj, clust_col = "cluster")
}
\arguments{
\item{clust_df}{Dataframe with context columns and a column with cluster
membership for that context.}

\item{dist_obj}{Distance object with sites in the same order as \code{clust_df}.}

\item{clust_col}{Character. Name of column containing cluster membership in
\code{clust_df}.}
}
\value{
\code{clust_df} with additional columns for silhouette width (as
\code{sil_width}) and neighbouring cluster (as \code{neighbour}).
}
\description{
Create a dataframe of silhouette widths for a clustering
}
