% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sil_plot.R
\name{make_sil_plot}
\alias{make_sil_plot}
\title{Make a silhouette plot from a silhouette dataframe}
\usage{
make_sil_plot(
  sil_df,
  clust_col = "cluster",
  clust_colours = NULL,
  include_labels = length(unique(sil_df[clust_col][[1]])) < 30
)
}
\arguments{
\item{sil_df}{Silhouette dataframe. See \code{make_sil_df()}.}

\item{clust_col}{Character name of column in \code{sil_df} containing the clusters.}

\item{clust_colours}{If null (default) \code{viridis::viridis} is used to colour
clusters. Otherwise a dataframe with clusters and their associated colours.
Needs to be the same length as levels in \code{clust_col} (or \code{levs}).}

\item{include_labels}{Logical. Include cluster labels on plot? Can get very
crowded if there are many clusters.}
}
\value{
ggplot
}
\description{
Make a silhouette plot from a silhouette dataframe
}
