% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary_df.R
\name{make_summary_df}
\alias{make_summary_df}
\title{Simple summary of a clustering.}
\usage{
make_summary_df(clust_df, clust_col = "cluster", min_sites = 10)
}
\arguments{
\item{clust_df}{Dataframe with context columns and a column with cluster
membership for that context.}

\item{clust_col}{Name of column in \code{clust_df} with class membership.}

\item{min_sites}{Desired minimum absolute number of sites in a class.}
}
\value{
single row tibble with summary information
}
\description{
Simple summary of a clustering.
}
