% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_wide_df}
\alias{make_wide_df}
\title{Make a wide (usually site * taxa) data frame}
\usage{
make_wide_df(
  bio_df,
  context,
  taxa_col = "taxa",
  num_col = "use_cover",
  num_col_NA = 0
)
}
\arguments{
\item{bio_df}{Dataframe containing the site and taxa data.}

\item{context}{Character. Name of columns defining context.}

\item{taxa_col}{Character. Name of column containing taxa.}

\item{num_col}{Character. Name of column containing numeric abundance data (
usually 'cover' for plants).}

\item{num_col_NA}{Value to use to fill NA in wide table.}
}
\value{

}
\description{
Make a wide (usually site * taxa) data frame
}
