% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{sites_with_indicator}
\alias{sites_with_indicator}
\title{Sites with indicator}
\usage{
sites_with_indicator(ind_val_df, clust_df, thresh = 0.05)
}
\arguments{
\item{ind_val_df}{Dataframe from make_indval_df}

\item{clust_df}{Dataframe with column of cluster membership.}

\item{thresh}{Numeric. p_val threshold for acceptance as indicator taxa.}
}
\value{
Numeric. Number of sites across clustering that are in a cluster with
at least one indicator taxa.
}
\description{
Sites with indicator
}
