% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wss.R
\name{calc_wss}
\alias{calc_wss}
\title{Calculate within cluster sum of squares}
\usage{
calc_wss(
  clust_df,
  dist_obj = NULL,
  dist_mat = NULL,
  clust_col = "clust",
  do_sample = FALSE
)
}
\arguments{
\item{clust_df}{Dataframe with column of cluster membership.}

\item{dist_obj}{Distance object from which clusters were generated.}

\item{dist_mat}{Result from \code{as.matrix(dist_obj)}, or equivalent. Required if
\code{dist_obj} is not supplied.}

\item{clust_col}{Character. Name of column in \code{clust_df} that contains the
clusters.}

\item{do_boot}{Logical. Sample \code{clust_col} (without replacement) before
calculation. As step in gap-statistic calculation.}
}
\value{
Dataframe of within group sum-of-squares for each cluster.
}
\description{
Calculate within cluster sum of squares
}
\keyword{internal}
