% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{cluster_summarise}
\alias{cluster_summarise}
\title{Filter summarised clusters}
\usage{
cluster_summarise(df, clustcol = "clust", minsites = 10)
}
\arguments{
\item{df}{Dataframe containing clusters as a list column.}

\item{clustcol}{Name of list column containing clusters.}

\item{minsites}{Desired minimum absolute number of sites in a cluster.}
}
\value{
Dataframe of filtered clusters.
}
\description{
Filter summarised clusters
}
