% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{cluster_summarise}
\alias{cluster_summarise}
\title{Simple summary of a clustering.}
\usage{
cluster_summarise(df, clust_col = "cluster", min_sites = 10)
}
\arguments{
\item{df}{Dataframe containing clusters as a list column.}

\item{clust_col}{Name of column in \code{df} with class membership.}

\item{min_sites}{Desired minimum absolute number of sites in a class.}
}
\value{
single row tibble with summary information
}
\description{
Simple summary of a clustering.
}
