% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_df.R
\name{diagnostic_df}
\alias{diagnostic_df}
\title{Use a set of (continuous) columns to choose a good set of candidate rows}
\usage{
diagnostic_df(
  df,
  df_with_diag = tibble::tibble(diagnostic = "av_clust_size", high_good = TRUE,
    clust_sum = TRUE),
  context = c("method", "groups"),
  diag_col = "clust_sum",
  summarise_method = median,
  group_col = "groups",
  min_groups = NULL,
  max_groups = NULL,
  top_thresh = 0.25,
  best_thresh = 5
)
}
\arguments{
\item{df}{Dataframe with columns over which to find good candidates.}

\item{df_with_diag}{Dataframe mapping the name of possible diagnostics to cases (columns) in which to use those diagnostic.}

\item{context}{Character. Name of columns in df that define context.}

\item{diag_col}{Character. Name of df_with_diag column to use in this instance.}

\item{summarise_method}{Character. Name of method to use in summarising if there is more than one row per context.}

\item{group_col}{Character. Optional. Name of column to filter on (min_groups, max_groups)}

\item{min_groups}{Numeric. Minimum number of clusters in a clustering.}

\item{max_groups}{Numeric. Maximum number of clusters in a clustering.}

\item{top_thresh}{Numeric specifying the proportion of rows considered 'top'.}

\item{best_thresh}{Numeric specifying the absolute number of rows considered 'best'.}
}
\value{

}
\description{
Use a set of (continuous) columns to choose a good set of candidate rows
}
