% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_plot.R
\name{diagnostic_plot}
\alias{diagnostic_plot}
\title{Plot the results from diagnostic_df}
\usage{
diagnostic_plot(diagnosticdf, label = "diagnostic", displayall = FALSE)
}
\arguments{
\item{diagnosticdf}{Dataframe with results from diagnostic_df.}

\item{label}{Character label for the diagnostics (choose another column from diagnosticDF).}

\item{displayall}{Display all diagnostics or only those used to select best.}
}
\value{

}
\description{
Plot the results from diagnostic_df
}
