% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_cluster_df}
\alias{make_cluster_df}
\title{Make a cluster data frame}
\usage{
make_cluster_df(raw_clusters, site_df, site_col = "site")
}
\arguments{
\item{raw_clusters}{Dataframe of cluster membership. Can be result of
make_clusters.}

\item{site_df}{Dataframe with 'site' ids in same order as rawclusters.}

\item{site_col}{Character. Name of column in sitedf with sites.}
}
\value{
Dataframe with 'site','clust' (numeric) and 'cluster' (character)
}
\description{
Make a cluster data frame
}
