% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_clusters}
\alias{make_clusters}
\title{Apply clustering algorithms}
\usage{
make_clusters(
  method,
  dist_obj,
  group_range = 2:100,
  clust_col = "clust",
  clust_col_out = "cluster",
  sites
)
}
\arguments{
\item{method}{What method to use to create clusters.}

\item{dist_obj}{Distance object, from, say, \code{vegan::vegdist()},
\code{stats::dist()} or \code{parallelDist::parDist()}.}

\item{clust_col}{Character. Name to give column representing each of \code{groups}}

\item{clust_col_out}{Character. Name to give column representing \code{groups} in
words.}

\item{groups}{Numeric vector indicating the range of groups within a
clustering.}
}
\value{
Dataframe with groups and methods columns and a list column of the
clustering for that number of groups and method as a tibble with one column
(numeric) 'clust_col' indicating group membership.
}
\description{
Apply clustering algorithms
}
