% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_clusters}
\alias{make_clusters}
\title{Apply clustering algorithms}
\usage{
make_clusters(
  bio_df,
  env_df = NULL,
  methods_df = tibble(method = "average"),
  context,
  taxa_col = "taxa",
  num_col = NULL,
  env_cols,
  groups = 2:100,
  cores = 1,
  clust_col_in = "clust",
  clust_col_out = "cluster"
)
}
\arguments{
\item{bio_df}{Dataframe. Needs the 'cols' outlined below.}

\item{env_df}{Dataframe (wide) of environmental variables per sitecol.}

\item{methods_df}{Dataframe. What methods to use to create clusters.}

\item{context}{Character. Name of columns in \code{bio_df} that define the context.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{num_col}{Character. Name of column with numeric data indicating
relative abundance of taxa.}

\item{env_cols}{Character. Name of columns in \code{env_df} that have the}

\item{groups}{Numeric vector indicating the range of groups within a
clustering.}

\item{cores}{Numeric. Number of cores to use where parallel options exist.}

\item{clust_col_in}{Character name of column of raw class labels.}

\item{clust_col_out}{Character name of column of character class labels.}
}
\value{
Dataframe with groups and methods columns and a list column of the
clustering for that number of groups and method as a tibble with one column
(numeric) 'clust' indicating group membership.
}
\description{
Apply clustering algorithms
}
