% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_clusters}
\alias{make_clusters}
\title{Apply clustering algorithms}
\usage{
make_clusters(
  df,
  env_df = NULL,
  methods_df = tibble(method = "average"),
  taxa_col = "taxa",
  site_col = "list",
  num_col = NULL,
  groups = 2:100
)
}
\arguments{
\item{df}{Dataframe. Needs the 'cols' outlined below.}

\item{env_df}{Dataframe (wide) of environmental variables per sitecol.}

\item{methods_df}{Dataframe. What methods to use to create clusters.}

\item{taxa_col}{Character. Name of column with taxa.}

\item{site_col}{Chracter. Name of column with 'site'.}

\item{num_col}{Character. Name of column with numeric data indicating
relative abundance of taxa.}

\item{groups}{Numeric vector indicating the range of groups within a
clustering.}
}
\value{
Dataframe with groups and methods columns and a list column of the
clustering for that number of groups and method as a tibble with one column
(numeric) 'clust' indicating group membership.
}
\description{
Apply clustering algorithms
}
