% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters_explore.R
\name{make_clusters_explore}
\alias{make_clusters_explore}
\title{Explore clusters}
\usage{
make_clusters_explore(
  clusters_df,
  cores = 1,
  save_results = TRUE,
  out_exp = if (save_results) tempdir(),
  obj_list
)
}
\arguments{
\item{clusters_df}{Dataframe with column 'clusters'.}

\item{cores}{How many cores to use in \code{\link[multidplyr:new_cluster]{multidplyr::new_cluster()}}?}

\item{save_results}{Logical. If true, results will be saved along the way. If
the file already exists, that result will not be recreated.}

\item{out_exp}{Directory into which results are saved.}

\item{obj_list}{List of objects required:
\describe{
\item{dist_flor}{\code{dist(flor_wide)}.}
\item{dist_env}{\code{dist(env_wide)}.}
\item{flor_wide}{Wide version of flor_tidy.}
\item{flor_tidy}{Cleaned data set of taxa observations.}
\item{visit_cols}{context.}
\item{p_thresh}{usually 0.05.}
\item{n_sites}{How many unique contexts in \code{flor_tidy}.}
\item{most_freq_prop_thresh}{Threshold for proportion of sites in a cluster
that must all contain at least one taxa in common.}
}}
}
\value{
Each output saved along the way to \code{out_exp}. Outputs are not made if
they have already been saved. Created outputs returned in list with elements:
\describe{
\item{clusters_sil}{floristics result from \code{\link[=make_sil_df]{make_sil_df()}}.}
\item{clusters_wss}{floristics result from \code{\link[=calc_wss]{calc_wss()}}.}
\item{clusters_sil_env}{env result from \code{\link[=make_sil_df]{make_sil_df()}}.}
\item{clusters_wss_env}{env result from \code{\link[=calc_wss]{calc_wss()}}.}
\item{clusters_ind_val}{indicator values for floristics from
\code{\link[=make_ind_val_df]{make_ind_val_df()}}, \code{\link[=clusters_with_indicator]{clusters_with_indicator()}} and
\code{\link[=sites_with_indicator]{sites_with_indicator()}}.}
\item{clusters_freq}{floristics results from \code{\link[=clusters_with_freq_taxa]{clusters_with_freq_taxa()}}
and \code{\link[=sites_with_freq_taxa]{sites_with_freq_taxa()}}}
}
}
\description{
Explore clusters using silhouette width, within-group sum-of-squares,
indicators and widespread taxa.
}
