% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ind_val_df.R
\name{make_ind_val_df}
\alias{make_ind_val_df}
\title{Create a dataframe of indicator values for a clustering}
\usage{
make_ind_val_df(
  clust_df = NULL,
  bio_df,
  context,
  clust_col = "cluster",
  cov_col = "use_cover",
  taxa_col = "taxa",
  ...
)
}
\arguments{
\item{clust_df}{Dataframe with context columns and a column with cluster
membership for that context. Optional if \code{clust_col} appears in bio_wide.}

\item{bio_df}{Dataframe containing the site and taxa data in long format.}

\item{context}{Character. Name(s) of column(s) that define the context.}

\item{clust_col}{Character. Name of column containing cluster membership.}

\item{cov_col}{Character. Name of column containing abundance data (often
'cover' values).}

\item{taxa_col}{Character. Name of column containing the taxa names.}

\item{...}{Passed to \code{labdsv::indval()}}
}
\value{
Dataframe of each taxa and the cluster (clust as numeric, cluster as
character) class to which it is most likely an indicator, plus the following
values from \code{labdsv::indval()} output: ind_val, p_val, abu and frq.
}
\description{
Create a dataframe of indicator values for a clustering
}
