% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_ind_val_df}
\alias{make_ind_val_df}
\title{Create a dataframe of indicator values}
\usage{
make_ind_val_df(clust_df, df_wide, clust_col = "clust")
}
\arguments{
\item{clust_df}{Dataframe including a column with cluster membership}

\item{df_wide}{Dataframe of sites * taxa. In the same order as clustdf.}

\item{clust_col}{Column containing cluster membership in clustdf as name or
index.}
}
\value{
Dataframe of each taxa and the cluster (clust as numeric, cluster as
character) class to which it is most likely an indicator, plus the following
values from labdsv::indval output: ind_val, p_val, abu and frq.
}
\description{
Create a dataframe of indicator values
}
