% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_neigh_text.R
\name{make_neigh_text}
\alias{make_neigh_text}
\title{Make text to describe silhouette.}
\usage{
make_neigh_text(cluster, sil_df)
}
\arguments{
\item{cluster}{Name of cluster.}

\item{sil_df}{Output from \code{make_sil_df()}}
}
\value{
Sentence for use in silhouette section of report.
}
\description{
Make text to describe silhouette.
}
