% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_sil_df}
\alias{make_sil_df}
\title{Create a dataframe of silhouette widths}
\usage{
make_sil_df(clustdf, distobj, clustcol = "clust")
}
\arguments{
\item{clustdf}{Dataframe with column of cluster membership.}

\item{distobj}{Distance object with sites in the same order as clustdf.}

\item{clustcol}{Name or index of column containing cluster membership in
clustdf as name or index.}
}
\value{
Dataframe of silhouette width per site, including neighbouring
cluster.
}
\description{
Create a dataframe of silhouette widths
}
