% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{make_sil_df}
\alias{make_sil_df}
\title{Create a dataframe of silhouette widths}
\usage{
make_sil_df(clust_df, dist_obj, clust_col = "clust")
}
\arguments{
\item{clust_df}{Dataframe with column of cluster membership.}

\item{dist_obj}{Distance object with sites in the same order as clustdf.}

\item{clust_col}{Name or index of column containing cluster membership in
\code{clust_df}.}
}
\value{
Dataframe of silhouette width per site, including neighbouring
cluster.
}
\description{
Create a dataframe of silhouette widths
}
