% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{sites_with_freq_taxa}
\alias{sites_with_freq_taxa}
\title{Title}
\usage{
sites_with_freq_taxa(
  clust_df,
  bio_df,
  clust_col = "clust",
  taxa_col = "taxa",
  site_col = "cell",
  thresh = 0.9
)
}
\arguments{
\item{clust_df}{Dataframe with column of cluster membership and join column to
\code{bio_df}}

\item{bio_df}{Dataframe with taxa}

\item{clust_col}{Character. Name of column in clustdf with clusters.}

\item{taxa_col}{Character. Name of column in \code{bio_df} with taxa.}

\item{site_col}{Character. Name of column in clustdf and \code{bio_df} with 'sites'.}

\item{thresh}{Numeric. Proportion of sites in a cluster at which >= 1 taxa
need to occur.}
}
\value{
Numeric. Number of sites across clustering that are in a cluster with
at least one taxa that occurs in greater than 'thresh' proportion of sites
in that cluster.
}
\description{
Title
}
