% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterings.R
\name{sites_with_indicator}
\alias{sites_with_indicator}
\title{Sites with indicator}
\usage{
sites_with_indicator(indvaldf, clustdf, thresh = 0.05)
}
\arguments{
\item{indvaldf}{Dataframe from make_indval_df}

\item{clustdf}{Dataframe with column of cluster membership.}

\item{thresh}{Numeric. pval threshold for acceptance as indicator taxa.}
}
\value{
Numeric. Number of sites across clustering that are in a cluster with
at least one indicator taxa.
}
\description{
Sites with indicator
}
